+++ example/lib/pages/map_inside_listview.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_map/flutter_map.dart';
 import 'package:flutter_map_example/misc/tile_providers.dart';
-import 'package:flutter_map_example/plugins/zoombuttons_plugin_option.dart';
+import 'package:flutter_map_example/plugins/zoombuttons_plugin.dart';
 import 'package:flutter_map_example/widgets/drawer/menu_drawer.dart';
 import 'package:latlong2/latlong.dart';
