+++ example/lib/plugins/scalebar_utils.dart
