+++ example/lib/plugins/zoombuttons_plugin.dart
